refactor(audit): move the connection emitter out of the middleware directory
`recordConnectionCreated` is a plain function called from the connectionCreated
hook, never mounted on a route, so `.middleware.ts` claimed something it isn't -
its two neighbours there export actual request handlers. It now sits next to its
only caller.
