+++ packages/server/lib/hooks/auditConnection.ts
