+++ packages/server/lib/hooks/hooks.ts
@@ -18,9 +18,9 @@ import {
 import { Err, getLogger, isHosted, Ok, report } from '@nangohq/utils';
 import { sendAuth as sendAuthWebhook } from '@nangohq/webhooks';
 
-import { recordConnectionCreated } from '../middleware/auditConnection.middleware.js';
 import { slackService } from '../services/slack.js';
 import { getOrchestrator } from '../utils/utils.js';
+import { recordConnectionCreated } from './auditConnection.js';
 import executeVerificationScript from './connection/credentials-verification-script.js';
 import { postConnectionCreation } from './connection/on/post-connection-creation.js';
 import postConnection from './connection/post-connection.js';
